attach `scanFiles` directly on the scan dir result
When a new file is discovered, we need to make sure that it actually
is allowed to be analyzed. To do this, we have to make sure that the
file matches the `globs` we know about.

We could scan the globs and pull in a library to verify that the file is
allowed.

Alternatively we could create a function in Rust that receives the files
and the list of globs. While this works, this means that we have to send
the globs every time we want to scan a new file.

Instead, this attaches a `scanFiles` directly to the `scanDirResult`
which means we can access the `globs` from scope in Rust without
sending it "over the wire".